Added documentation and menu items
updated the readme, and added Help and Development menu items, which are just children of the PopUp class. The text these messages show is defined within its corresponding widget class, so in main.py all you see is the call to initiate the message window and not a huge mound of text.