Load implemented, other features too
Got load to work. Also added buttons so you can shuffle between members
as you create them. You can then edit a member that you're on, and it
will automatically update any joint that member is a part of. This makes
it easy to load a design, and make small changes. Also added some
documentation improvements here and there, and fixed directory issues.