Added results tab, fix small things
Results tab displays the results. Fixed popups as there was an error in
there. A couple small math mistakes with box beams, I shouldn't rely on
them being a perfect box (width=thickness) but this is the best way to
make use of the box. Also made it so hole supports come in as ints not
floats.
I also added another one of those class property decorators to simplify
button enabling, and added the ability to delete the joints list and
start over. However, to delete the members list you must start over.
I also just noticed that the pin bending doesn't work with any design,
it was specific to my design in class, so I will have to think about
what to do there.