Added design components to the second tab
Wanted to put something else on the results tab so I changed it to the
'Components and Results' tab, and added a design components textbox to
display updated members and joints. For the most part, this is updated
when mIndex and jIndex are updated, which is convenient. But when you
edit a member, mIndex and jIndex are not altered, so I had to force an
update there.
When Calculate or View Components is clicked, it switches tabs (and
still logs the results in the console). Finally I added a basic title
font to the constants file.