Fix small bug with filenames, documented stuff
A weird bug I had was with loading and saving filepaths. I set the initial directory for the load and save dialogs to the correct one, checked with print statements, using os.path.dirname(os.path.abspath(__file__)) to get the directory. But sometimes it put me in the wrong directory, so I switched from abspath() to realpath() and it works so far. I'm on the train away from internet when commiting this to my local master so I can't search up the difference between realpath() and abspath() :laughing:
Also updated documentation stuff.